feat: migrate to Next.js 16 and replace middleware with proxy

+++ app/(private)/scheduled-payments/_components/ScheduledPaymentsList.tsx
@@ -87,6 +87,7 @@ function ScheduledPaymentsList({
         onClose={() => setDeleteData(null)}
         confirmationQusestion={td("deleteScheduledPaymentMessage")}
         onDecision={() => {
+          if (!deleteData) return;
           onDelete([deleteData!._id]);
         }}
         loading={isPending}
